pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"


[project]
name = "yolo-train-pipeline"
version = "0.1.0"
description = "Generic dataset conversion, preparation, training, and ONNX export/optimization pipeline for Ultralytics YOLO"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "ultralytics>=8.4.19,<9",
  "pyyaml>=6.0",
  "tqdm>=4.66",
  "opencv-python>=4.10,<5",
  "matplotlib>=3.8,<4",
  "pillow>=10,<13",
  "gdown>=5.2,<6",
  "onnx>=1.20,<2",
  "onnxslim>=0.1.87",
]


[project.optional-dependencies]
cpu = [
  "onnxruntime>=1.24.3,<2",
]

cuda = [
  "onnxruntime-gpu>=1.24.3,<2",
  "onnxconverter-common>=1.16,<2",
]


[project.scripts]
yolo-convert-dataset = "tools.convert_dataset_to_yolo:main"
yolo-print-stats = "tools.print_yolo_dataset_stats:main"
yolo-prepare-dataset = "tools.prepare_yolo_dataset:main"
yolo-train = "tools.train:main"
yolo-report-ap = "tools.report_ap:main"

yolo-onnx-export = "tools.onnx.export:main"
yolo-onnx-optimize = "tools.onnx.optimize:main"
yolo-onnx-pipeline = "tools.onnx.pipeline:main"
yolo-benchmark-report = "tools.yolo_bench_report:main"


[tool.setuptools]
include-package-data = false


[tool.setuptools.packages.find]
include = ["core*", "tools*"]
exclude = ["data*", "models*", "reports*", "runs*", "configs*", "scripts*"]
